chore: split ui-tests/Util.ts into ui-tests/utils/ and fix broken imports
- Delete ui-tests/Util.ts; extract into 7 focused modules under ui-tests/utils/:
  constants.ts, editor.ts, extension.ts, settings.ts, terminal.ts,
  tree-view.ts, workbench.ts
- Update all ui-tests consumers to import from the new modules
- Fix MavenRuntimeDetector and StepsOnSaveManager: import from ./path and
  ./version instead of the now-deleted ./helpers
- Fix test file imports in src/test/views/ and src/test/extension/
  (depth was one level too shallow after Sub-Task 6 restructure)
- Fix OpenApiImportService.test.ts __dirname-relative stubs path
  (now at src/test/services/, needs one extra ../ to reach project root)
- Fix ImportOpenApiCommand.ts: import from OpenApiImportService
  (stale reference to old kebab-case filename)
