+++ src/test/services/OpenApiImportService.test.ts
@@ -751,7 +751,7 @@ suite('Pet Store OpenAPI Integration', function () {
 	});
 
 	test('should parse pet-store.openapi.json and generate REST DSL + Routes', async function () {
-		const petStoreJson = await fs.readFile(path.join(__dirname, '../../../../src/test/stubs', 'pet-store.openapi.json'), 'utf-8');
+		const petStoreJson = await fs.readFile(path.join(__dirname, '../../../../../src/test/stubs', 'pet-store.openapi.json'), 'utf-8');
 
 		const result = await service.parseOpenApi(petStoreJson, {
 			shouldGenerateRest: true,
