+++ src/test/views/deployments/DeploymentsProvider.test.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 import { assert } from 'chai';
-import { DeploymentsProvider, PortFileKey } from '../../views/deployments/DeploymentsProvider';
-import { PortManager } from '../../utils/PortManager';
-import { CamelTask, CamelTaskDefinition } from '../../tasks/CamelTask';
+import { DeploymentsProvider, PortFileKey } from '../../../views/deployments/DeploymentsProvider';
+import { PortManager } from '../../../utils/PortManager';
+import { CamelTask, CamelTaskDefinition } from '../../../tasks/CamelTask';
 
 // ─── Helpers ────────────────────────────────────────────────────────────────
