+++ src/utils/MavenRuntimeDetector.ts
@@ -5,7 +5,8 @@ import { KaotoOutputChannel } from '../extension/KaotoOutputChannel';
 import { KaotoCatalogService } from '../services/KaotoCatalogService';
 import { CamelExecutorFactory } from '../executors/CamelExecutorFactory';
 import { CamelLauncherExecutor } from '../executors/CamelLauncherExecutor';
-import { findFolderOfPomXml, normalizeVersionForSemver } from './helpers';
+import { findFolderOfPomXml } from './path';
+import { normalizeVersionForSemver } from './version';
 import { DEFAULT_CAMEL_VERSION_FALLBACK } from '../constants';
 
 /**
