+++ ui-tests/Z_IntegrationsViewNewProject.test.ts
@@ -31,14 +31,8 @@ import {
 	VSBrowser,
 	WebDriver,
 } from 'vscode-extension-tester';
-import {
-	expandFolderItemsInTreeStructuredView,
-	expandViews,
-	getKaotoViewControl,
-	getTreeItem,
-	getTreeItemActionButton,
-	openResourcesAndWaitForActivation,
-} from './Util';
+import { openResourcesAndWaitForActivation } from './utils/extension';
+import { expandFolderItemsInTreeStructuredView, expandViews, getKaotoViewControl, getTreeItem, getTreeItemActionButton } from './utils/tree-view';
 
 /**
  * This test needs to be always executed as last in suite
